Extract STL viewer logic into module
