+++ frontend/scripts.js
@@ -43,17 +43,6 @@ const COLOR_THEMES = Object.freeze({
   LIGHT: 'light',
   DARK: 'dark',
 });
-const STL_DEFAULT_DIMENSIONS = Object.freeze({
-  width: 4,
-  depth: 6,
-  height: 3,
-});
-const INCH_TO_MM = 25.4;
-const THREE_CDN_SOURCES = Object.freeze([
-  'https://cdn.jsdelivr.net/npm/three@0.161.0/build/three.min.js',
-  'https://unpkg.com/three@0.161.0/build/three.min.js',
-]);
-let threeLoaderPromise = null;
 const OVERLAY_COORDINATE_SCALE = 1000;
 
 const HINT_TUNING_INPUT_IDS = Object.freeze({
@@ -128,24 +117,6 @@ export function bootGridFinium(dependencies = {}) {
   setupHintTuningControls();
   setupThemeToggle();
 
-  const stlDesignerOptions = {
-    viewerId: 'stl-viewer',
-    widthInputId: 'stl-width',
-    depthInputId: 'stl-depth',
-    heightInputId: 'stl-height',
-    summaryId: 'stl-summary',
-    downloadButtonId: 'stl-download',
-    resetButtonId: 'stl-reset',
-  };
-
-  ensureThreeJs()
-    .then(() => {
-      initStlDesigner(stlDesignerOptions);
-    })
-    .catch((error) => {
-      console.error('GridFinium: unable to load Three.js from any CDN source. Falling back to canvas renderer.', error);
-      initStlDesigner(stlDesignerOptions);
-    });
 }
 
 function setupThemeToggle() {
@@ -699,76 +670,6 @@ function waitForOpenCv() {
   });
 }
 
-function ensureThreeJs() {
-  if (typeof THREE !== 'undefined') return Promise.resolve();
-  if (threeLoaderPromise) return threeLoaderPromise;
-
-  threeLoaderPromise = new Promise((resolve, reject) => {
-    const trySource = (index) => {
-      if (typeof THREE !== 'undefined') {
-        resolve();
-        return;
-      }
-
-      if (index >= THREE_CDN_SOURCES.length) {
-        reject(new Error('Three.js failed to load'));
-        return;
-      }
-
-      const script = document.createElement('script');
-      script.src = THREE_CDN_SOURCES[index];
-      script.async = true;
-      script.crossOrigin = 'anonymous';
-
-      const cleanup = () => {
-        script.removeEventListener('load', handleLoad);
-        script.removeEventListener('error', handleError);
-      };
-
-      const handleLoad = () => {
-        cleanup();
-        if (typeof THREE !== 'undefined') {
-          resolve();
-        } else {
-          script.remove();
-          trySource(index + 1);
-        }
-      };
-
-      const handleError = () => {
-        cleanup();
-        script.remove();
-        trySource(index + 1);
-      };
-
-      script.addEventListener('load', handleLoad);
-      script.addEventListener('error', handleError);
-      document.head.appendChild(script);
-    };
-
-    trySource(0);
-  }).then(
-    () => {
-      return undefined;
-    },
-    (error) => {
-      threeLoaderPromise = null;
-      throw error;
-    },
-  );
-
-  return threeLoaderPromise;
-}
-
-function initStlDesigner(options) {
-  if (typeof THREE !== 'undefined') {
-    return initThreeStlDesigner(options);
-  }
-
-  return initCanvasStlDesigner(options);
-}
-
-
 function createStepRenderer(container, options = {}) {
   ensureProcessingStyles();
 
@@ -1294,770 +1195,4 @@ function ensureProcessingStyles() {
   `;
   document.head.appendChild(style);
 }
-
-function initThreeStlDesigner({
-  viewerId,
-  widthInputId,
-  depthInputId,
-  heightInputId,
-  summaryId,
-  downloadButtonId,
-  resetButtonId,
-}) {
-  const viewerRoot = document.getElementById(viewerId);
-  const widthInput = document.getElementById(widthInputId);
-  const depthInput = document.getElementById(depthInputId);
-  const heightInput = document.getElementById(heightInputId);
-  const summaryNode = document.getElementById(summaryId);
-  const downloadButton = document.getElementById(downloadButtonId);
-  const resetButton = document.getElementById(resetButtonId);
-
-  if (!viewerRoot || !widthInput || !depthInput || !heightInput || !summaryNode) return null;
-  if (typeof THREE === 'undefined') {
-    viewerRoot.textContent = 'Three.js failed to load, so the 3D preview is unavailable.';
-    return null;
-  }
-
-  const renderer = new THREE.WebGLRenderer({ antialias: true, alpha: true });
-  renderer.setPixelRatio(window.devicePixelRatio);
-  const initialWidth = Math.max(
-    1,
-    viewerRoot.clientWidth || viewerRoot.offsetWidth || viewerRoot.scrollWidth || 480,
-  );
-  const initialHeight = Math.max(
-    1,
-    viewerRoot.clientHeight || viewerRoot.offsetHeight || viewerRoot.scrollHeight || 320,
-  );
-  const initialAspect = initialWidth / initialHeight;
-  renderer.setSize(initialWidth, initialHeight);
-  renderer.domElement.style.width = '100%';
-  renderer.domElement.style.height = '100%';
-  renderer.domElement.style.cursor = 'grab';
-  renderer.domElement.style.touchAction = 'none';
-  viewerRoot.appendChild(renderer.domElement);
-
-  const scene = new THREE.Scene();
-  const camera = new THREE.PerspectiveCamera(42, initialAspect, 0.1, 2000);
-  camera.position.set(12, 8, 12);
-  camera.lookAt(0, 0, 0);
-
-  scene.add(new THREE.AmbientLight(0xffffff, 0.6));
-
-  const keyLight = new THREE.DirectionalLight(0xffffff, 0.85);
-  keyLight.position.set(10, 12, 18);
-  scene.add(keyLight);
-
-  const rimLight = new THREE.DirectionalLight(0xffffff, 0.35);
-  rimLight.position.set(-8, -6, -10);
-  scene.add(rimLight);
-
-  const ground = new THREE.GridHelper(40, 10, 0xa5b4fc, 0xe0e7ff);
-  ground.position.y = -STL_DEFAULT_DIMENSIONS.height / 2;
-  scene.add(ground);
-
-  const modelGroup = new THREE.Group();
-  modelGroup.rotation.set(Math.PI / 10, Math.PI / 8, 0);
-  scene.add(modelGroup);
-
-  let activeMesh = null;
-
-  const ensureCameraFrame = (dimensions) => {
-    const maxDimension = Math.max(dimensions.width, dimensions.depth, dimensions.height);
-    const distance = Math.max(12, maxDimension * 3.2);
-    camera.position.set(distance, distance * 0.75, distance);
-    camera.near = Math.max(0.1, maxDimension / 20);
-    camera.far = distance * 12;
-    camera.lookAt(0, 0, 0);
-    camera.updateProjectionMatrix();
-  };
-
-  const rebuildMesh = (dimensions) => {
-    if (activeMesh) {
-      modelGroup.remove(activeMesh);
-      activeMesh.geometry.dispose();
-      activeMesh.material.dispose();
-    }
-
-    const geometry = new THREE.BoxGeometry(dimensions.width, dimensions.height, dimensions.depth);
-    const material = new THREE.MeshStandardMaterial({
-      color: 0x4f46e5,
-      metalness: 0.08,
-      roughness: 0.4,
-    });
-
-    activeMesh = new THREE.Mesh(geometry, material);
-    modelGroup.add(activeMesh);
-
-    ground.position.y = -dimensions.height / 2;
-    ensureCameraFrame(dimensions);
-  };
-
-  const resizeRenderer = () => {
-    const { clientWidth, clientHeight } = viewerRoot;
-    if (clientWidth === 0 || clientHeight === 0) return;
-    renderer.setSize(clientWidth, clientHeight, false);
-    camera.aspect = clientWidth / clientHeight;
-    camera.updateProjectionMatrix();
-  };
-
-  const dragState = {
-    active: false,
-    pointerId: null,
-    previous: new THREE.Vector2(),
-  };
-
-  const handlePointerDown = (event) => {
-    dragState.active = true;
-    dragState.pointerId = event.pointerId;
-    dragState.previous.set(event.clientX, event.clientY);
-    renderer.domElement.setPointerCapture(event.pointerId);
-    renderer.domElement.style.cursor = 'grabbing';
-  };
-
-  const handlePointerMove = (event) => {
-    if (!dragState.active || dragState.pointerId !== event.pointerId) return;
-    const deltaX = event.clientX - dragState.previous.x;
-    const deltaY = event.clientY - dragState.previous.y;
-
-    modelGroup.rotation.y += deltaX * 0.005;
-    modelGroup.rotation.x += deltaY * 0.005;
-    modelGroup.rotation.x = Math.max(-Math.PI / 2, Math.min(Math.PI / 2, modelGroup.rotation.x));
-
-    dragState.previous.set(event.clientX, event.clientY);
-  };
-
-  const releasePointer = (event) => {
-    if (dragState.pointerId !== event.pointerId) return;
-    dragState.active = false;
-    dragState.pointerId = null;
-    renderer.domElement.releasePointerCapture(event.pointerId);
-    renderer.domElement.style.cursor = 'grab';
-  };
-
-  const handleWheel = (event) => {
-    event.preventDefault();
-    const zoomFactor = Math.exp(event.deltaY * 0.001);
-    camera.position.multiplyScalar(zoomFactor);
-  };
-
-  renderer.domElement.addEventListener('pointerdown', handlePointerDown);
-  renderer.domElement.addEventListener('pointermove', handlePointerMove);
-  renderer.domElement.addEventListener('pointerup', releasePointer);
-  renderer.domElement.addEventListener('pointercancel', releasePointer);
-  renderer.domElement.addEventListener('pointerleave', () => {
-    dragState.active = false;
-    dragState.pointerId = null;
-    renderer.domElement.style.cursor = 'grab';
-  });
-  renderer.domElement.addEventListener('wheel', handleWheel, { passive: false });
-  window.addEventListener('resize', resizeRenderer);
-
-  const controls = configureStlControls({
-    widthInput,
-    depthInput,
-    heightInput,
-    summaryNode,
-    downloadButton,
-    resetButton,
-    onChange: (dimensions) => {
-      rebuildMesh(dimensions);
-    },
-  });
-
-  resizeRenderer();
-
-  const animate = () => {
-    requestAnimationFrame(animate);
-    renderer.render(scene, camera);
-  };
-
-  animate();
-
-  return {
-    getDimensions: () => controls.getDimensions(),
-    reset: () => controls.reset(),
-  };
-}
-
-function initCanvasStlDesigner({
-  viewerId,
-  widthInputId,
-  depthInputId,
-  heightInputId,
-  summaryId,
-  downloadButtonId,
-  resetButtonId,
-}) {
-  const viewerRoot = document.getElementById(viewerId);
-  const widthInput = document.getElementById(widthInputId);
-  const depthInput = document.getElementById(depthInputId);
-  const heightInput = document.getElementById(heightInputId);
-  const summaryNode = document.getElementById(summaryId);
-  const downloadButton = document.getElementById(downloadButtonId);
-  const resetButton = document.getElementById(resetButtonId);
-
-  if (!viewerRoot || !widthInput || !depthInput || !heightInput || !summaryNode) return null;
-
-  const canvas = document.createElement('canvas');
-  canvas.style.width = '100%';
-  canvas.style.height = '100%';
-  canvas.style.display = 'block';
-  canvas.style.touchAction = 'none';
-  canvas.classList.add('stl-viewer__canvas');
-
-  viewerRoot.replaceChildren(canvas);
-
-  const context = canvas.getContext('2d');
-  if (!context) {
-    viewerRoot.textContent = 'Your browser does not support the canvas 3D preview.';
-    return null;
-  }
-
-  const faces = [
-    [0, 1, 2, 3],
-    [4, 5, 6, 7],
-    [3, 2, 6, 7],
-    [0, 1, 5, 4],
-    [1, 2, 6, 5],
-    [0, 3, 7, 4],
-  ];
-
-  const edges = [
-    [0, 1],
-    [1, 2],
-    [2, 3],
-    [3, 0],
-    [4, 5],
-    [5, 6],
-    [6, 7],
-    [7, 4],
-    [0, 4],
-    [1, 5],
-    [2, 6],
-    [3, 7],
-  ];
-
-  const lightDirection = normalizeVector({ x: 0.6, y: 0.85, z: 1 });
-  const state = {
-    rotationX: Math.PI / 10,
-    rotationY: Math.PI / 8,
-    zoom: 1,
-  };
-
-  const displaySize = {
-    width: Math.max(1, viewerRoot.clientWidth || viewerRoot.offsetWidth || viewerRoot.scrollWidth || 480),
-    height: Math.max(1, viewerRoot.clientHeight || viewerRoot.offsetHeight || viewerRoot.scrollHeight || 320),
-    dpr: window.devicePixelRatio || 1,
-  };
-
-  let latestDimensions = { ...STL_DEFAULT_DIMENSIONS };
-  let pendingFrame = null;
-
-  const ensureCanvasSize = () => {
-    displaySize.width = Math.max(
-      1,
-      viewerRoot.clientWidth || viewerRoot.offsetWidth || viewerRoot.scrollWidth || displaySize.width,
-    );
-    displaySize.height = Math.max(
-      1,
-      viewerRoot.clientHeight || viewerRoot.offsetHeight || viewerRoot.scrollHeight || displaySize.height,
-    );
-    displaySize.dpr = window.devicePixelRatio || 1;
-
-    canvas.width = Math.max(1, Math.round(displaySize.width * displaySize.dpr));
-    canvas.height = Math.max(1, Math.round(displaySize.height * displaySize.dpr));
-    canvas.style.width = `${displaySize.width}px`;
-    canvas.style.height = `${displaySize.height}px`;
-
-    scheduleRender();
-  };
-
-  const scheduleRender = () => {
-    if (pendingFrame) return;
-    pendingFrame = window.requestAnimationFrame(() => {
-      pendingFrame = null;
-      render();
-    });
-  };
-
-  const render = () => {
-    const width = displaySize.width;
-    const height = displaySize.height;
-    if (width === 0 || height === 0) return;
-
-    context.setTransform(displaySize.dpr, 0, 0, displaySize.dpr, 0, 0);
-    context.clearRect(0, 0, width, height);
-
-    const gradient = context.createLinearGradient(0, 0, 0, height);
-    gradient.addColorStop(0, '#eef2ff');
-    gradient.addColorStop(1, '#e2e8f0');
-    context.fillStyle = gradient;
-    context.fillRect(0, 0, width, height);
-
-    const dims = latestDimensions;
-    const halfWidth = width / 2;
-    const halfHeight = height / 2;
-    const maxDimension = Math.max(dims.width, dims.height, dims.depth);
-    const minCanvasSize = Math.min(width, height);
-    const scale = (minCanvasSize * 0.5) / (maxDimension || 1);
-    const baseDistance = (maxDimension || 1) * scale * 3.2;
-    const distance = baseDistance / state.zoom;
-    const focalLength = distance;
-
-    const baseVertices = buildBoxVertices(dims, scale);
-    const rotatedVertices = baseVertices.map((vertex) => rotateVertex(vertex, state.rotationX, state.rotationY));
-    const projectedVertices = rotatedVertices.map((vertex) => projectVertex(vertex, halfWidth, halfHeight, distance, focalLength));
-
-    drawShadow(context, dims, maxDimension, minCanvasSize, halfWidth, halfHeight);
-
-    const visibleFaces = faces
-      .map((indices) => buildFaceData(indices, rotatedVertices, projectedVertices, distance))
-      .filter((face) => face && face.viewDot < 0)
-      .map((face) => ({
-        ...face,
-        fill: shadeColor({ r: 79, g: 70, b: 229 }, lightDirection, face.normal),
-      }));
-
-    visibleFaces.sort((a, b) => b.averageDepth - a.averageDepth);
-
-    visibleFaces.forEach((face) => {
-      context.beginPath();
-      face.projected.forEach((point, index) => {
-        if (index === 0) {
-          context.moveTo(point.x, point.y);
-        } else {
-          context.lineTo(point.x, point.y);
-        }
-      });
-      context.closePath();
-      context.fillStyle = face.fill;
-      context.fill();
-    });
-
-    context.beginPath();
-    edges.forEach(([startIndex, endIndex]) => {
-      const start = projectedVertices[startIndex];
-      const end = projectedVertices[endIndex];
-      context.moveTo(start.x, start.y);
-      context.lineTo(end.x, end.y);
-    });
-    context.lineWidth = 1.2;
-    context.strokeStyle = 'rgba(30, 41, 59, 0.35)';
-    context.stroke();
-  };
-
-  const dragState = {
-    active: false,
-    pointerId: null,
-    previous: { x: 0, y: 0 },
-  };
-
-  canvas.addEventListener('pointerdown', (event) => {
-    dragState.active = true;
-    dragState.pointerId = event.pointerId;
-    dragState.previous.x = event.clientX;
-    dragState.previous.y = event.clientY;
-    canvas.setPointerCapture(event.pointerId);
-    canvas.style.cursor = 'grabbing';
-  });
-
-  canvas.addEventListener('pointermove', (event) => {
-    if (!dragState.active || dragState.pointerId !== event.pointerId) return;
-    const deltaX = event.clientX - dragState.previous.x;
-    const deltaY = event.clientY - dragState.previous.y;
-    dragState.previous.x = event.clientX;
-    dragState.previous.y = event.clientY;
-
-    state.rotationY += deltaX * 0.005;
-    state.rotationX += deltaY * 0.005;
-    state.rotationX = Math.max(-Math.PI / 2, Math.min(Math.PI / 2, state.rotationX));
-
-    scheduleRender();
-  });
-
-  const releasePointer = (event) => {
-    if (dragState.pointerId !== event.pointerId) return;
-    dragState.active = false;
-    dragState.pointerId = null;
-    canvas.releasePointerCapture(event.pointerId);
-    canvas.style.cursor = 'grab';
-  };
-
-  canvas.addEventListener('pointerup', releasePointer);
-  canvas.addEventListener('pointercancel', releasePointer);
-  canvas.addEventListener('pointerleave', () => {
-    dragState.active = false;
-    dragState.pointerId = null;
-    canvas.style.cursor = 'grab';
-  });
-
-  canvas.addEventListener(
-    'wheel',
-    (event) => {
-      event.preventDefault();
-      const zoomFactor = Math.exp(event.deltaY * 0.001);
-      state.zoom = Math.min(6, Math.max(0.4, state.zoom / zoomFactor));
-      scheduleRender();
-    },
-    { passive: false },
-  );
-
-  const controls = configureStlControls({
-    widthInput,
-    depthInput,
-    heightInput,
-    summaryNode,
-    downloadButton,
-    resetButton,
-    onChange: (dimensions) => {
-      latestDimensions = { ...dimensions };
-      scheduleRender();
-    },
-  });
-
-  window.addEventListener('resize', ensureCanvasSize);
-  ensureCanvasSize();
-
-  return {
-    getDimensions: () => controls.getDimensions(),
-    reset: () => controls.reset(),
-  };
-}
-
-function configureStlControls({
-  widthInput,
-  depthInput,
-  heightInput,
-  summaryNode,
-  downloadButton,
-  resetButton,
-  onChange,
-}) {
-  let currentDimensions = { ...STL_DEFAULT_DIMENSIONS };
-
-  const formatDimensionValue = (value) => {
-    const normalized = Math.round(value * 1000) / 1000;
-    return Number.isFinite(normalized) ? normalized.toString() : '';
-  };
-
-  const applyDimensions = (dimensions, { commitInputs = false } = {}) => {
-    currentDimensions = { ...dimensions };
-
-    if (commitInputs) {
-      widthInput.value = formatDimensionValue(currentDimensions.width);
-      depthInput.value = formatDimensionValue(currentDimensions.depth);
-      heightInput.value = formatDimensionValue(currentDimensions.height);
-    }
-
-    updateStlSummary(summaryNode, currentDimensions);
-    if (typeof onChange === 'function') {
-      onChange({ ...currentDimensions });
-    }
-  };
-
-  const readDimension = (input) => {
-    const value = Number.parseFloat(input.value);
-    if (!Number.isFinite(value) || value <= 0) return null;
-    return value;
-  };
-
-  const handleLiveInput = () => {
-    const next = { ...currentDimensions };
-    let hasChange = false;
-
-    const width = readDimension(widthInput);
-    if (width !== null && width !== next.width) {
-      next.width = width;
-      hasChange = true;
-    }
-
-    const depth = readDimension(depthInput);
-    if (depth !== null && depth !== next.depth) {
-      next.depth = depth;
-      hasChange = true;
-    }
-
-    const height = readDimension(heightInput);
-    if (height !== null && height !== next.height) {
-      next.height = height;
-      hasChange = true;
-    }
-
-    if (hasChange) {
-      applyDimensions(next);
-    }
-  };
-
-  const commitDimensions = () => {
-    const width = readDimension(widthInput) ?? currentDimensions.width;
-    const depth = readDimension(depthInput) ?? currentDimensions.depth;
-    const height = readDimension(heightInput) ?? currentDimensions.height;
-    applyDimensions({ width, depth, height }, { commitInputs: true });
-  };
-
-  const resetDimensions = () => {
-    applyDimensions({ ...STL_DEFAULT_DIMENSIONS }, { commitInputs: true });
-  };
-
-  const handleDownload = () => {
-    const stlContent = generateBoxStl(currentDimensions);
-    const fileName = buildStlFileName(currentDimensions);
-    const blob = new Blob([stlContent], { type: 'model/stl' });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = fileName;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-  };
-
-  widthInput.addEventListener('input', handleLiveInput);
-  depthInput.addEventListener('input', handleLiveInput);
-  heightInput.addEventListener('input', handleLiveInput);
-
-  widthInput.addEventListener('change', commitDimensions);
-  depthInput.addEventListener('change', commitDimensions);
-  heightInput.addEventListener('change', commitDimensions);
-
-  if (downloadButton) {
-    downloadButton.addEventListener('click', (event) => {
-      event.preventDefault();
-      handleDownload();
-    });
-  }
-
-  if (resetButton) {
-    resetButton.addEventListener('click', (event) => {
-      event.preventDefault();
-      resetDimensions();
-    });
-  }
-
-  resetDimensions();
-
-  return {
-    getDimensions: () => ({ ...currentDimensions }),
-    reset: resetDimensions,
-  };
-}
-
-function updateStlSummary(summaryNode, dimensions) {
-  if (!summaryNode) return;
-  const { width, depth, height } = dimensions;
-  const format = (value) => Math.round(value * 100) / 100;
-  summaryNode.textContent = `Box dimensions: ${format(width)}" × ${format(depth)}" × ${format(height)}".`;
-}
-
-function buildBoxVertices(dimensions, scale) {
-  const hx = ((dimensions.width || 0) * scale) / 2;
-  const hy = ((dimensions.height || 0) * scale) / 2;
-  const hz = ((dimensions.depth || 0) * scale) / 2;
-
-  return [
-    { x: -hx, y: -hy, z: -hz },
-    { x: hx, y: -hy, z: -hz },
-    { x: hx, y: hy, z: -hz },
-    { x: -hx, y: hy, z: -hz },
-    { x: -hx, y: -hy, z: hz },
-    { x: hx, y: -hy, z: hz },
-    { x: hx, y: hy, z: hz },
-    { x: -hx, y: hy, z: hz },
-  ];
-}
-
-function rotateVertex(vertex, rotationX, rotationY) {
-  const cosY = Math.cos(rotationY);
-  const sinY = Math.sin(rotationY);
-  const cosX = Math.cos(rotationX);
-  const sinX = Math.sin(rotationX);
-
-  const x1 = vertex.x * cosY - vertex.z * sinY;
-  const z1 = vertex.x * sinY + vertex.z * cosY;
-  const y1 = vertex.y * cosX - z1 * sinX;
-  const z2 = vertex.y * sinX + z1 * cosX;
-
-  return { x: x1, y: y1, z: z2 };
-}
-
-function projectVertex(vertex, halfWidth, halfHeight, distance, focalLength) {
-  const z = vertex.z + distance;
-  const perspective = focalLength / (z || 1);
-
-  return {
-    x: halfWidth + vertex.x * perspective,
-    y: halfHeight - vertex.y * perspective,
-    depth: z,
-  };
-}
-
-function drawShadow(context, dimensions, maxDimension, minCanvasSize, centerX, centerY) {
-  const safeMax = maxDimension || 1;
-  const scaleWidth = (dimensions.width / safeMax) * minCanvasSize * 0.28;
-  const scaleDepth = (dimensions.depth / safeMax) * minCanvasSize * 0.22;
-  const offsetY = (dimensions.height / safeMax) * minCanvasSize * 0.14;
-
-  context.save();
-  context.fillStyle = 'rgba(15, 23, 42, 0.12)';
-  context.beginPath();
-  context.ellipse(
-    centerX,
-    centerY + offsetY,
-    Math.max(12, scaleWidth),
-    Math.max(8, scaleDepth),
-    0,
-    0,
-    Math.PI * 2,
-  );
-  context.fill();
-  context.restore();
-}
-
-function buildFaceData(indices, rotatedVertices, projectedVertices, cameraDistance) {
-  const rotated = indices.map((index) => rotatedVertices[index]);
-  const projected = indices.map((index) => projectedVertices[index]);
-
-  const edgeA = subtractVector(rotated[1], rotated[0]);
-  const edgeB = subtractVector(rotated[2], rotated[0]);
-  const normal = crossVector(edgeA, edgeB);
-  const normalLength = Math.hypot(normal.x, normal.y, normal.z);
-  if (!normalLength) return null;
-  const unitNormal = {
-    x: normal.x / normalLength,
-    y: normal.y / normalLength,
-    z: normal.z / normalLength,
-  };
-
-  const center = rotated.reduce(
-    (acc, point) => ({ x: acc.x + point.x, y: acc.y + point.y, z: acc.z + point.z }),
-    { x: 0, y: 0, z: 0 },
-  );
-  center.x /= rotated.length;
-  center.y /= rotated.length;
-  center.z /= rotated.length;
-
-  const centerCamera = {
-    x: center.x,
-    y: center.y,
-    z: center.z + cameraDistance,
-  };
-
-  const toCamera = {
-    x: -centerCamera.x,
-    y: -centerCamera.y,
-    z: -centerCamera.z,
-  };
-  const toCameraLength = Math.hypot(toCamera.x, toCamera.y, toCamera.z) || 1;
-  const unitToCamera = {
-    x: toCamera.x / toCameraLength,
-    y: toCamera.y / toCameraLength,
-    z: toCamera.z / toCameraLength,
-  };
-
-  const averageDepth = projected.reduce((sum, point) => sum + point.depth, 0) / projected.length;
-
-  return {
-    normal: unitNormal,
-    projected,
-    averageDepth,
-    viewDot: dotVector(unitNormal, unitToCamera),
-  };
-}
-
-function shadeColor(baseColor, lightDirection, normal) {
-  const dot = Math.max(0, dotVector(lightDirection, normal));
-  const intensity = 0.35 + 0.65 * dot;
-
-  const r = Math.round(Math.min(255, baseColor.r * intensity));
-  const g = Math.round(Math.min(255, baseColor.g * intensity));
-  const b = Math.round(Math.min(255, baseColor.b * intensity));
-
-  return `rgb(${r}, ${g}, ${b})`;
-}
-
-function normalizeVector(vector) {
-  const length = Math.hypot(vector.x, vector.y, vector.z) || 1;
-  return {
-    x: vector.x / length,
-    y: vector.y / length,
-    z: vector.z / length,
-  };
-}
-
-function subtractVector(a, b) {
-  return { x: a.x - b.x, y: a.y - b.y, z: a.z - b.z };
-}
-
-function crossVector(a, b) {
-  return {
-    x: a.y * b.z - a.z * b.y,
-    y: a.z * b.x - a.x * b.z,
-    z: a.x * b.y - a.y * b.x,
-  };
-}
-
-function dotVector(a, b) {
-  return a.x * b.x + a.y * b.y + a.z * b.z;
-}
-
-function generateBoxStl(dimensions) {
-  const width = dimensions.width * INCH_TO_MM;
-  const depth = dimensions.depth * INCH_TO_MM;
-  const height = dimensions.height * INCH_TO_MM;
-
-  const hx = width / 2;
-  const hy = height / 2;
-  const hz = depth / 2;
-
-  const vertices = [
-    [-hx, -hy, -hz],
-    [hx, -hy, -hz],
-    [hx, hy, -hz],
-    [-hx, hy, -hz],
-    [-hx, -hy, hz],
-    [hx, -hy, hz],
-    [hx, hy, hz],
-    [-hx, hy, hz],
-  ];
-
-  const facets = [
-    { normal: [0, 0, -1], triangles: [[0, 1, 2], [0, 2, 3]] },
-    { normal: [0, 0, 1], triangles: [[4, 5, 6], [4, 6, 7]] },
-    { normal: [0, 1, 0], triangles: [[3, 2, 6], [3, 6, 7]] },
-    { normal: [0, -1, 0], triangles: [[0, 1, 5], [0, 5, 4]] },
-    { normal: [1, 0, 0], triangles: [[1, 6, 2], [1, 5, 6]] },
-    { normal: [-1, 0, 0], triangles: [[0, 7, 3], [0, 4, 7]] },
-  ];
-
-  const lines = ['solid gridfinium_box'];
-
-  facets.forEach(({ normal, triangles }) => {
-    const normalLine = `  facet normal ${normal.map((value) => value.toFixed(6)).join(' ')}`;
-    triangles.forEach((triangle) => {
-      lines.push(normalLine, '    outer loop');
-      triangle.forEach((index) => {
-        const vertex = vertices[index];
-        lines.push(`      vertex ${vertex.map((value) => value.toFixed(6)).join(' ')}`);
-      });
-      lines.push('    endloop', '  endfacet');
-    });
-  });
-
-  lines.push('endsolid gridfinium_box');
-  return `${lines.join('\n')}\n`;
-}
-
-function buildStlFileName(dimensions) {
-  const parts = [dimensions.width, dimensions.depth, dimensions.height].map((value) => {
-    const rounded = Math.round(value * 100) / 100;
-    return String(rounded).replace(/\./g, '_');
-  });
-  return `gridfinium-box-${parts.join('x')}.stl`;
-}
-
-
-
-
-
